1. Fixed a regression bug that broke ssh for older OS with requiretty set in sudo
2. Added a command line option to use the devel version of the
gemini repository code. This was needed to maintain the stable version that works and
enable us to modify the backend code for make it work with exogeni.
What this means is simple terms is that the code pulled from gemini.netlab.uky.edu onto the nodes
will be from the devel version if --devel is added to the list of args on the init and instrumentize scripts
So the same python scripts will work for both version.